[feat] #114 지난 & 다가올 데이트 일정 전체 조회 API 연동

+++ app/src/main/java/org/sopt/dateroad/domain/model/Date.kt
@@ -3,7 +3,7 @@ package org.sopt.dateroad.domain.model
 import org.sopt.dateroad.presentation.type.DateTagType
 
 data class Date(
-    val dateId: Int,
+    val dateId: Long,
     val title: String,
     val date: String,
     val city: String,
